Semana14Hackaton/jquispe/requirements.txt
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.3
ctx==0.1.2
Django==3.1.1
django-braces==1.14.0
django-suit==0.2.28
html5lib==1.1
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.5
pylint==2.6.0
PyPDF2==1.26.0
pytz==2020.1
reportlab==3.5.49
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
Unipath==1.1
webencodings==0.5.1
wrapt==1.12.1
xhtml2pdf==0.2.4
